# openapi.yml
openapi: 3.0.3
info:
  title: HMS Backend API
  version: 1.0.0
  description: |
    OpenAPI specification based solely on the application's actual Express routes and validation rules.
  contact:
    name: HMS Admin
    email: mrfortrancer@gmail.com
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://hms.fortrancer.in
    description: Production server
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request payload or missing required information.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication is required or token is invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Caller does not have the required permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          example: Something went wrong.
        statusCode:
          type: integer
          example: 500
        code:
          type: string
          example: INTERNAL_SERVER_ERROR
        details:
          type: object
          additionalProperties: true
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
          example: Operation completed successfully.
...
    MedicalRecordCreateRequest:
      type: object
      required:
        - doctorId
        - appointmentId
        - patientId
        - diagnosis
        - complaint
        - symptoms
        - createdBy
      properties:
        doctorId:
          type: string
          example: EMP-2001
        appointmentId:
          type: string
          example: APT-9001
        patientId:
          type: string
          example: PAT-1001
        diagnosis:
          type: string
          example: Hypertension
        complaint:
          type: string
          example: Headache and dizziness
        symptoms:
          type: string
          example: Elevated blood pressure and fatigue
        medications:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: Amlodipine
              dosage:
                type: string
                example: 5mg
              frequency:
                type: string
                example: Once daily
              duration:
                type: string
                example: 30 days
        medicalObservations:
          type: array
          items:
            type: object
            properties:
              metricName:
                type: string
                example: Blood Pressure
              metricValue:
                type: string
                example: 140/90
              recordedTime:
                type: string
                format: date-time
                example: 2026-07-08T10:30:00.000Z
        notes:
          type: string
          example: Patient advised dietary changes and routine monitoring
        status:
          type: string
          example: Completed
        createdBy:
          type: string
          example: EMP-1001
    MedicalRecordUpdateRequest:
      type: object
      required:
        - medicalRecordId
        - doctorId
        - appointmentId
        - patientId
        - diagnosis
        - complaint
        - symptoms
        - createdBy
      properties:
        medicalRecordId:
          type: string
          example: REC-1001
        doctorId:
          type: string
          example: EMP-2001
        appointmentId:
          type: string
          example: APT-9001
        patientId:
          type: string
          example: PAT-1001
        diagnosis:
          type: string
          example: Hypertension
        complaint:
          type: string
          example: Headache and dizziness
        symptoms:
          type: string
          example: Elevated blood pressure and fatigue
        medications:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: Amlodipine
              dosage:
                type: string
                example: 5mg
              frequency:
                type: string
                example: Once daily
              duration:
                type: string
                example: 30 days
        medicalObservations:
          type: array
          items:
            type: object
            properties:
              metricName:
                type: string
                example: Blood Pressure
              metricValue:
                type: string
                example: 140/90
              recordedTime:
                type: string
                format: date-time
                example: 2026-07-08T10:30:00.000Z
        notes:
          type: string
          example: Updated patient history notes
        status:
          type: string
          example: Completed
        createdBy:
          type: string
          example: EMP-1001
        updatedBy:
          type: string
          example: EMP-1002
        updatedAt:
          type: string
          format: date-time
          example: 2026-07-09T12:00:00.000Z
    MedicalRecordQueryResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        total:
          type: integer
          example: 12
        page:
          type: integer
          example: 1
        totalPages:
          type: integer
          example: 3
    NodeCreateRequest:
      type: object
      required:
        - order
        - name
        - path
        - role
        - icon
      properties:
        order:
          type: integer
          example: 1
        name:
          type: string
          example: Dashboard
        path:
          type: string
          example: /dashboard
        role:
          type: array
          items:
            type: string
          example: ["Admin", "Doctor"]
        icon:
          type: string
          example: dashboard
    NodeEditRequest:
      type: object
      required:
        - nodeName
      properties:
        nodeName:
          type: string
          example: Dashboard
        data:
          type: object
          additionalProperties: true
    NodeDeleteRequest:
      type: object
      required:
        - nodeName
      properties:
        nodeName:
          type: string
          example: Dashboard
...
paths:
...
  /node/getNodes:
    get:
      tags: [Node]
      summary: Get nodes by role
      description: Returns UI nodes permitted for the requested role.
      operationId: getNodes
      parameters:
        - in: query
          name: role
          required: true
          schema:
            type: string
            example: Doctor
      responses:
        '200':
          description: Nodes returned successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NodeCreateRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /node/createNode:
    post:
      tags: [Node]
      summary: Create node
      description: Creates a new sidebar/navigation node.
      operationId: createNode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeCreateRequest'
      responses:
        '201':
          description: Node created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeCreateRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /node/editNode:
    post:
      tags: [Node]
      summary: Edit node
      description: Updates a node by nodeName.
      operationId: editNode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeEditRequest'
      responses:
        '200':
          description: Node edited successfully.
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /node/deleteNode:
    post:
      tags: [Node]
      summary: Delete node
      description: Deletes a node by nodeName.
      operationId: deleteNode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeDeleteRequest'
      responses:
        '200':
          description: Node deleted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /medicalRecord/createMedicalRecord:
    post:
      tags: [MedicalRecord]
      summary: Create medical record
      description: Creates a medical record for an appointment.
      operationId: createMedicalRecord
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicalRecordCreateRequest'
      responses:
        '201':
          description: Medical record created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /medicalRecord/updateMedicalRecord:
    post:
      tags: [MedicalRecord]
      summary: Update medical record
      description: Updates an existing medical record.
      operationId: updateMedicalRecord
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicalRecordUpdateRequest'
      responses:
        '200':
          description: Medical record updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /medicalRecord/getMedicalRecordStats:
    get:
      tags: [MedicalRecord]
      summary: Get medical record stats
      description: Returns aggregated medical record counts.
      operationId: getMedicalRecordStats
      responses:
        '200':
          description: Statistics returned successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  medicalRecordCount:
                    type: integer
                  completedCount:
                    type: integer
                  draftCount:
                    type: integer
                  deletedCount:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /medicalRecord/getMedicalRecords:
    get:
      tags: [MedicalRecord]
      summary: Get medical records
      description: Returns paginated medical records.
      operationId: getMedicalRecords
      parameters:
        - in: query
          name: page
          required: true
          schema:
            type: integer
            example: 1
        - in: query
          name: limit
          required: true
          schema:
            type: integer
            example: 10
        - in: query
          name: patientId
          required: false
          schema:
            type: string
            example: PAT-1001
        - in: query
          name: doctorId
          required: false
          schema:
            type: string
            example: EMP-2001
        - in: query
          name: isClientApp
          required: false
          schema:
            type: boolean
            example: false
        - in: query
          name: selectedText
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Medical records returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicalRecordQueryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /medicalRecord/getMedicalRecordById:
    get:
      tags: [MedicalRecord]
      summary: Get medical record by ID
      description: Returns a medical record for the supplied medicalRecordId.
      operationId: getMedicalRecordById
      parameters:
        - in: query
          name: medicalRecordId
          required: true
          schema:
            type: string
            example: REC-1001
      responses:
        '200':
          description: Medical record returned successfully.
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /medicalRecord/deleteMedicalRecord:
    post:
      tags: [MedicalRecord]
      summary: Delete medical record
      description: Marks a medical record as deleted.
      operationId: deleteMedicalRecord
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - medicalRecordId
              properties:
                medicalRecordId:
                  type: string
                  example: REC-1001
      responses:
        '200':
          description: Medical record deleted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'